pyproject.toml
[project]
name = "tms-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.4",
    "fastapi>=0.119.0",
    "folium>=0.20.0",
    "from-root>=1.3.0",
    "geopy>=2.4.1",
    "ipykernel>=7.0.0",
    "matplotlib>=3.10.7",
    "mysqlclient>=2.2.7",
    "numpy>=2.3.3",
    "ortools>=9.14.6206",
    "pandas>=2.3.3",
    "pydantic>=2.12.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]
